requirements.txt
Flask~=3.0.3
SQLAlchemy~=2.0.32
Werkzeug~=3.0.3
python-dotenv~=1.0.1